Add take picture and select picture function
Be aware, there is a strange error. Every time I save a change, The error will jump up says " _reactNative.Linking.removeEventListener is not a function (it is undefined)".
But this error will disappear when I save something again, even it's just a space.
Might be render or navigation problem.